% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batnets-data.R
\docType{data}
\name{batnets}
\alias{batnets}
\title{A dataset of adjacency matrices of bat-insect interactions}
\format{A nested list of adjacancy matrices}
\usage{
batnets
}
\description{
Data from metabarcoding guano samples of insectivorous bats,
using the Zeale (2011) primers. The data is organised as a nested list,
with the first list level being the clustering level used, the second list level
being the dataset in question, and the final objects being interaction matrixes.
The data is from Big Bend National Park, Texas and the SAFE project, Malaysia.
}
\references{
A large-scale forest fragmentation experiment: the Stability of Altered Forest Ecosystems Project
Robert M. Ewers, Raphael K. Didham, Lenore Fahrig, Gonçalo Ferraz, Andy Hector, Robert D. Holt, Valerie Kapos, Glen Reynolds, Waidi Sinun, Jake L. Snaddon, Edgar C. Turner
Phil. Trans. R. Soc. B 2011 366 3292-3302; DOI: 10.1098/rstb.2011.0049. Published 17 October 2011
}
\keyword{datasets}
