% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_scripts.R
\name{deviance.phylogenetic}
\alias{deviance.phylogenetic}
\title{Deviance in Phylogenetic Specificity to Null Models}
\usage{
\method{deviance}{phylogenetic}(
  x,
  utree,
  null.model = c("taxa.labels", "richness", "frequency", "sample.pool",
    "phylogeny.pool", "independentswap", "trialswap"),
  iterations = 100,
  abundance.weighted = TRUE,
  trim = TRUE,
  notify = TRUE
)
}
\arguments{
\item{x}{Data frame. Host x symbiont data frame with hosts populating rows and symbionts populating columns.}

\item{utree}{Newick formatted phylogenetic tree.}

\item{null.model}{Randomization method. Usage borrowed directly from picante::mpd 
Specify as "taxa.labels", "richness", "frequency", "sample.pool", "phylogeny.pool", "independentswap" or "trialswap".}

\item{abundance.weighted}{Logical. TRUE calculates abundance-weighted mean pairwise phylogenetic distance. 
FALSE calculates presence-absence mean pairwise phylogenetic distance.}

\item{trim}{Logical. TRUE removes symbionts that occupy one host sample. FALSE keeps all symbionts.}

\item{notify}{Logical. TRUE prints the current iteration of the for loop.}
}
\value{
A data frame with columns that refer to the host sample identifiers, mean deviance in phylogenetic specificity, 
standard error of structural specificities, number of symbionts per host sample and average symbiont read abundance. This function does not specify 
a model argument whereas the other three deviance- functions do. Phylogenetic specificity...
}
\description{
Calculate the deviance in observed phylogenetic specificity to a null model of structural specificity per symbiont. 
Deviance calculations are measured per symbiont and averaged per host sample. For example, all symbionts within a given host 
are evaluated for their host specificity across the entire host community. The host specificities of each symbiont are averaged
to calculate the mean host specificity for symbionts within a given host.
}
\examples{
# Calculate mean deviance per symbiont per host sample and visualize null vs. observed host specifities 
deviance.phylogenetic(quad.rarefied, utree, null.model = "taxa.labels", iterations = 100, abundance.weighted = TRUE, trim = TRUE, notify = TRUE)
}
