% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metcalcs.R
\name{metcalcs}
\alias{metcalcs}
\title{calculate variance in network level metrics caused by MOTU clustering level}
\usage{
metcalcs(networks, indices, network_level = "both",
  list_format = "clust_net")
}
\arguments{
\item{networks}{A list of networks}

\item{indices}{A vector of indices to be calculated. See the functions networklevel and computeModules in package bipartite for details}

\item{network_level}{The network level to analyse}

\item{list_format}{The input must be a list of networks. It can be a nested list with the levels either being clustering, then network identity ('clust_net'), or network identity,
then clustering ('net_clust'), or alternatively a list of only one network origin, with the names of the list items corresponding to the clustering levels used ('clust_only').}
}
\value{
Produces a dataframe showing which metrics are robust in your dataset to clustering-level effects
}
\description{
Acts as a wrapper for bipartite's networklevel and computeModules functions, calculating the difference between your given network-level metrics for a series of networks and clustering thresholds
}
\examples{
metcalcs(networks= batnets, indices =c('functional complementarity','web asymmetry', 'Fisher alpha'), network_level = 'higher')
}
\seealso{
\code{\link{line_plot}} for visualisation of the resulting data
}
