% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_scripts.R
\name{network.specificity}
\alias{network.specificity}
\title{Network Specificity}
\usage{
network.specificity(x, abundance.weighted = TRUE, trim = TRUE)
}
\arguments{
\item{x}{Data frame. Host x symbiont data frame with hosts populating rows and symbionts populating columns.}

\item{abundance.weighted}{Logical. TRUE calculates Shannon's H per symbiont. FALSE calculates host richness per symbiont.}

\item{trim}{Logical. TRUE removes symbionts that occupy one host sample. FALSE keeps all symbionts.}
}
\value{
A data frame with symbiont identifiers and network specificity values.
}
\description{
Calculate network specificity not corrected by null mdoels.
}
\examples{
# Calculate host richness per symbiont
network.specificity(quad.rarefied, abundance.weighted = TRUE, trim = TRUE)

# Calculate Shannon's H per symbiont
network.specificity(quad.rarefied, abundance.weighted = TRUE, trim = TRUE)
}
