% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_scripts.R
\name{null.beta}
\alias{null.beta}
\title{Beta-Specificity Null Models}
\usage{
null.beta(
  x,
  index = c("morisita.horn", "horn", "sorensen"),
  randomization.method = c("r2dtable", "swap.web", "vaznull", "shuffle.web", "mgen"),
  iterations = 100,
  trim = TRUE,
  notify = TRUE
)
}
\arguments{
\item{x}{Data frame of hosts populating rows and symbionts populating columns.}

\item{index}{Character. Method for calculation with the Morisita-Horn, Horn or Sorensen Indices.}

\item{randomization.method}{Randomization method. Usage borrowed directly from bipartite::nullmodel. 
Specify as "r2dtable", "swap.web", "vaznull", "shuffle.web" or "mgen".}

\item{iterations}{Integer. Indicate the number of randomized communities to generate.}

\item{trim}{Logical. TRUE removes symbionts that occupy one host sample. FALSE keeps all symbionts.}

\item{notify}{Logical. TRUE prints the current iteration of the for loop. NOTE: This function can take some time.}
}
\value{
A data frame with columns that refer to symbiont identifiers, absolute read abundance, beta-specificities and randomization
identifiers.
}
\description{
Generate null models and calculate beta-specificity per symbiont within each community randomization.
}
\examples{
# Generate randomized communities and calculate beta-specificity per symbiont 
null.beta(quad.rarefied, index = "morisita.horn", randomization.method = "shuffle.web", iterations = 100, trim = TRUE, notify = TRUE)
}
