% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_scripts.R
\name{null.structural}
\alias{null.structural}
\title{Structural Specificity Null Models}
\usage{
null.structural(
  x,
  iterations = 100,
  abundance.weighted = TRUE,
  randomization.method = c("r2dtable", "swap.web", "vaznull", "shuffle.web", "mgen"),
  trim = TRUE,
  notify = TRUE
)
}
\arguments{
\item{x}{Data frame of hosts populating rows and symbionts populating columns.}

\item{iterations}{Integer. Indicate the number of randomized communities to generate.}

\item{abundance.weighted}{Logical. TRUE calculates Shannon's H per symbiont. FALSE calculates host richness per symbiont.}

\item{randomization.method}{Randomization method. Usage borrowed directly from bipartite::nullmodel. 
Specify as "r2dtable", "swap.web", "vaznull", "shuffle.web" or "mgen".}

\item{trim}{Logical. TRUE removes symbionts that occupy one host sample. FALSE keeps all symbionts.}

\item{notify}{Logical. TRUE prints the current iteration of the for loop.}
}
\value{
A data frame with columns that refer to symbiont identifiers, absolute read abundance, structural specificities and randomization
identifiers.
}
\description{
Generate null models and calculate structural specificity per symbiont within each community randomization.
}
\examples{
# Generate randomized communities and calculate structural specificity per symbiont 
null.structural(quad.rarefied, iterations = 100, abundance.weighted = TRUE, randomization.method = "shuffle.web", trim = TRUE, notify = TRUE)
}
