% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultrametric_tree.R
\docType{data}
\name{phylo.tree}
\alias{phylo.tree}
\title{Ultrametric tree for phylogenetic specificity analyses}
\format{
A phylogenetic tree 26 tips in newick format.
}
\source{
Jin Y, Qian H. 2019. V.PhyloMaker: an R package that can generate very large phylogenies for vascular plants. Ecography. 42:1353–1359
}
\usage{
data(phylo.tree)
}
\description{
Phylogenetic tree used to calculate phylogenetic specificity was pruned from a backbone phylogeny representing all extant vascular plant families in North America with the ‘phylo.maker’ function in the V.PhyloMaker package (Jin and Qian 2019). This ultrametric tree includes 26 plant samples from 13 plant species. Plant conspecific phylogenetic distances were manually changed to zero.
}
\examples{
data(phylo.tree)

}
\references{
Small, E. 1976. Insect pollinators of the Mer Bleue peat bog of Ottawa. Canadian Field Naturalist 90:22-28
}
\keyword{datasets}
