% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_scripts.R
\name{phylogenetic.specificity}
\alias{phylogenetic.specificity}
\title{Phylogenetic Specificity}
\usage{
phylogenetic.specificity(x, utree, abundance.weighted = TRUE, trim = TRUE)
}
\arguments{
\item{x}{Data frame. Host x symbiont data frame with hosts populating rows and symbionts populating columns.}

\item{utree}{Newick formatted phylogenetic tree.}

\item{abundance.weighted}{Logical. TRUE calculates abundance-weighted mean pairwise phylogenetic distance. 
FALSE calculates presence-absence mean pairwise phylogenetic distance.}

\item{trim}{Logical. TRUE removes symbionts that occupy one host sample. FALSE keeps all symbionts.}
}
\value{
A data frame with symbiont identifiers and phylogenetic specificity values.
}
\description{
Calculate phylogenetic specificity not corrected by null mdoels.
}
\examples{
# Calculate mean pairwise phylogenetic distance per symbiont
phylogenetic.specificity(quad.rarefied, utree, abundance.weighted = TRUE, trim = TRUE)
}
