% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_matrix.R
\docType{data}
\name{quad.rarefied}
\alias{quad.rarefied}
\title{Community matrix for host specificity analyses}
\format{
A data frame
}
\source{
Apigo and Oono 2020.
}
\usage{
data(quad.rarefied)
}
\description{
Data from a endophyte survey at the James San Jacinto Reserve on 
host specificity between foliar fungal endophytes and plants. 
This matrix includes 80 plant samples from 38 plant species and 
1 117 endophyte amplicon sequence variants.
}
\examples{
data(quad.rarefied)

\donttest{structural.specificity(quad.rarefied)}
}
\references{
Apigo and Oono 2020.
}
\keyword{datasets}
