% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomized_ranges.R
\name{randomized_ranges}
\alias{randomized_ranges}
\title{Identify meaningful trends emerging from MOTU clustering thresholds}
\usage{
randomized_ranges(networks, input_format = "clust_net", indices,
  network_level = "both", n_perm = 1000, sums_to_preserve = "both",
  summarise = T, quantiles_to_return = c(0.025, 0.975),
  out_format = "data.frame", actual_vals = F)
}
\arguments{
\item{networks}{A nested list of networks}

\item{input_format}{The input must be a list of networks, either a single network identity generated at different clustering levels ('clust_only'),
or a nested list of networks, with the levels either being clustering, then network identity ('clust_net'), or network identity,
then clustering ('net_clust').}

\item{indices}{A vector of indices to be calculated. See the functions networklevel and computeModules in package bipartite for details}

\item{network_level}{The network level to analyse}

\item{n_perm}{The number of permutations to run}

\item{sums_to_preserve}{The sums to be preserved in randomisation. Possible values are 'none', 'rows', 'columns', or 'both'}

\item{summarise}{Should the randomised values be summarised to the upper and lower quantiles, or should all randomised values be returned?}

\item{quantiles_to_return}{The quantiles desired if summarising the output}

\item{out_format}{The format for the data to be output in. Either a dataframe ('data.frame') or list ('list')}

\item{actual_vals}{Should the actual values for each network be calculated?}
}
\value{
Produces either a dataframe or list for your desired metrics and randomisations
}
\description{
Acts as a wrapper for the \code{\link{metcalcs}} function, randomising a networks across a series of clustering levels
}
\seealso{
\code{\link{metcalcs}}
}
