% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_scripts.R
\name{relative.phylogenetic}
\alias{relative.phylogenetic}
\title{Deviance in Phylogenetic Specificity to Null Models}
\usage{
relative.phylogenetic(
  x,
  utree,
  null.model = c("taxa.labels", "richness", "frequency", "sample.pool",
    "phylogeny.pool", "independentswap", "trialswap"),
  iterations = 100,
  abundance.weighted = TRUE,
  model = c("first", "second"),
  trim = TRUE,
  notify = TRUE
)
}
\arguments{
\item{x}{Data frame. Host x symbiont data frame with hosts populating rows and symbionts populating columns.}

\item{utree}{Newick formatted host phylogenetic tree. Cophenetic distances will be calculated the function.}

\item{null.model}{Randomization method. Usage borrowed directly from picante::mpd 
Specify as "taxa.labels", "richness", "frequency", "sample.pool", "phylogeny.pool", "independentswap" or "trialswap".}

\item{iterations}{Integer. Indicate the number of randomized communities to generate.}

\item{abundance.weighted}{Logical. TRUE calculates abundance-weighted mean pairwise phylogenetic distance. 
FALSE calculates presence-absence mean pairwise phylogenetic distance.}

\item{model}{Character. Specify whether the null expectation should be approximated as a first-(linear) or second-(quadratic) order function.}

\item{trim}{Logical. TRUE removes symbionts that occupy one host species. FALSE keeps all symbionts.}

\item{notify}{Logical. TRUE prints the current iteration of the for loop.}
}
\value{
A data frame with columns that refer to the host sample identifiers, mean deviance in phylogenetic specificity, 
standard error of structural specificities, number of symbionts per host sample and average symbiont read abundance.
}
\description{
Calculate the deviance in observed phylogenetic specificity to a null model of structural specificity per symbiont.
}
\details{
Please make sure your host or sample identifiers match the labels on the host phylogenetic tree. 

Host specificity is calculated for a symbiont across the entire host community. For example, a symbiont found in given host would be evaluated for its host specificity relative to all hosts that are present in a given dataset. 

Deviance calculations are measured per symbiont and averaged per host sample. The host specificities of each symbiont are averaged to calculate the mean host specificity for symbionts within a given host. Unlike relative structural and network specificity, relative phylogenetic specificity is not calculated as the vertical distance between absolute host specificity and the null model with respect to read abundance. Instead, relative phylogenetic specificity is calculated as the difference between absolute phylogenetic specificity and the mean value of phylogenetic specificity from randomized communities divided by the standard deviation of the null distribution of absolute phylogenetic specificity. 

A relative host specificity value greater than zero indicates that an endophyte was more host-specific relative to endophytes with the same read abundances within randomized communities. 

see documentation for `picante` for more documentation regarding the ses.mpd function
}
\examples{
# Calculate mean deviance per symbiont per host sample
\donttest{phy.dev <- relative.phylogenetic(quad.rarefied, utree, null.model = "taxa.labels", iterations = 100)}
\donttest{phy.dev[[1]]} # View data frame of output 
# Visualize null vs. absolute host specifities 
\donttest{phy.dev[[2]]} # View first graph 
\donttest{phy.dev[[81]]} # View last graph 
}
