% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_scripts.R
\name{structural.specificity}
\alias{structural.specificity}
\title{Structural Specificity}
\usage{
structural.specificity(x, abundance.weighted = TRUE, trim = TRUE)
}
\arguments{
\item{x}{Data frame of hosts populating rows and symbionts populating columns.}

\item{abundance.weighted}{Logical. TRUE calculates Shannon's H per symbiont. FALSE calculates host richness per symbiont.}

\item{trim}{Logical. TRUE removes symbionts that occupy one host sample. FALSE keeps all symbionts.}
}
\value{
A data frame with symbiont identifiers and structural specificity values.
}
\description{
Calculate structural specificity not corrected by null mdoels.
}
\examples{
# Calculate host richness per symbiont
structural.specificity(quad.rarefied, abundance.weighted = TRUE, trim = TRUE)

# Calculate Shannon's H per symbiont
structural.specificity(quad.rarefied, abundance.weighted = TRUE, trim = TRUE)
}
