% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultrametric_tree.R
\docType{data}
\name{utree}
\alias{utree}
\title{Ultrametric tree for phylogenetic specificity analyses}
\format{
A phylogenetic tree 80 tips in newick format.
}
\source{
Apigo and Oono 2020
}
\usage{
data(utree)
}
\description{
Phylogenetic tree used to calculate phylogenetic specificity was pruned from a backbone phylogeny representing all extant vascular plant families in North America with the ‘phylo.maker’ function in the V.PhyloMaker package (Jin and Qian 2019). Phylogenetic distances among plant conspecifics were manually changed to zero such that phylogenetic distance between conspecifics did not contribute to measurements of phylogenetic specificity. This ultrametric tree includes 80 plant samples from 38 plant species.
}
\examples{
data(utree)

}
\references{
Apigo and Oono 2020.
}
\keyword{datasets}
