% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_scripts.R
\name{beta.specificity}
\alias{beta.specificity}
\title{Structural Specificity}
\usage{
beta.specificity(
  x,
  index = c("morisita.horn", "horn", "sorensen"),
  trim = TRUE,
  notify = TRUE
)
}
\arguments{
\item{x}{Data frame of hosts populating rows and symbionts populating columns.}

\item{index}{Character. Method for calculation with the Morisita-Horn, Horn or Sorensen Indices.}

\item{trim}{Logical. TRUE removes symbionts that occupy one host sample. FALSE keeps all symbionts.}
}
\value{
A data frame with symbiont identifiers and beta-specificity values.
}
\description{
Calculate structural specificity not corrected by null mdoels.
}
\examples{
# Calculate beta-specificity
beta.specificity(quad.rarefied, index = "morisita.horn", trim = TRUE, notify = TRUE)
}
