% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{line_plot}
\alias{line_plot}
\title{Returns a figure showing how conclusions could change over MOTU clustering thresholds. Nb this function assumes that all datasets analysed using metcalcs were generated with the exact same clustering thresholds}
\usage{
line_plot(input, network, clustering, metric, value, plotname = NULL,
  colour = F, palette)
}
\arguments{
\item{input}{An input data frame, as output fully-formatted by the metcalcs function}

\item{network}{The network column of the data frame}

\item{clustering}{The clustering level column of the data frame}

\item{metric}{The column of the data frame containing the metrics desired for analysis}

\item{value}{The column of the data frame containing the values obtained for each metric}

\item{plotname}{A title for the plot, defaults to none}

\item{colour}{Should the plot be in colour?}

\item{palette}{A palette of colours to plot}
}
\value{
Produces a simple plot showing which metrics are robust in your dataset to clustering-level effects
}
\description{
Returns a figure showing how conclusions could change over MOTU clustering thresholds. Nb this function assumes that all datasets analysed using metcalcs were generated with the exact same clustering thresholds
}
\examples{
m <- metcalcs(networks= batnets, indices =c('functional complementarity','web asymmetry', 'Fisher alpha'), network_level = 'higher')
line_plot(input = m, metric = 'metric', network = 'network', clustering = 'clustering', value = 'value', plotname = 'Sabah dataset')
}
\seealso{
\code{\link{metcalcs}} which this function visualises the output of
}
