% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOTUS.R
\docType{package}
\name{LOTUS-package}
\alias{LOTUS}
\alias{LOTUS-package}
\title{LOTUS: Showing the effects of clustering on bipartite network data}
\description{
Checks how MOTU clustering thresholds alter network conclusions in bipartite networks
}
\details{
LOTUS currently consists of two basic commands: \code{\link{metcalcs}} for comparing how network-level metrics vary
over MOTU clustering thresholds, and \code{\link{line_plot}} for visualising the results.
\code{\link{batnets}} is included as an example dataset.
See the vignette for details
}
\author{
\strong{Maintainer}: David Hemprich-Bennett \email{hemprich.bennett@gmail.com}

}
\keyword{internal}
