% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_scripts.R
\name{network.specificity}
\alias{network.specificity}
\title{Network Specificity}
\usage{
network.specificity(x, abundance.weighted = TRUE, trim = TRUE)
}
\arguments{
\item{x}{Data frame. Host and symbiont data frame with hosts populating rows and symbionts populating columns.}

\item{abundance.weighted}{Logical. TRUE calculates the Paired Difference Index per symbiont. FALSE calculates the Resource Range Index per symbiont.}

\item{trim}{Logical. TRUE removes symbionts that occupy one host species. FALSE keeps all symbionts.}
}
\value{
A data frame with symbiont identifiers and network specificity values.
}
\description{
Calculate absolute network specificity not relativized by null models.
}
\details{
Hosts are labeled by their species name with a period and number identifier (e.g., hostA.1) to differentiate host samples that are of the same species. This naming scheme is required because host specificity is quantified at the level of host species and not host samples.  If this naming scheme does not apply to your experimental design, you should still add in identifiers (e.g., .1, .2, .3, etc.) after each host species or sample identifer.

Host specificity for a symbiont is evaluated across the entire host community. More positive values indicate a narrower symbiont niche and higher host specificity.
}
\examples{
# Calculate Resource Range Index per symbiont
\donttest{rri <- network.specificity(quad.rarefied, abundance.weighted = FALSE, trim = TRUE)}
}
\references{
Austen Apigo and Ryoko Oono. 2021. Novel metrics reveal plant abundance, but not plant evolutionary history, shape host specificity in foliar fungal symbionts. In review.
}
