% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_scripts.R
\name{phylogenetic.specificity}
\alias{phylogenetic.specificity}
\title{Absolute Phylogenetic Specificity}
\usage{
phylogenetic.specificity(x, utree, abundance.weighted = TRUE, trim = TRUE)
}
\arguments{
\item{x}{Data frame. Host x symbiont data frame with hosts populating rows and symbionts populating columns.}

\item{utree}{Newick formatted host phylogenetic tree. Cophenetic distances will be calculated the function.}

\item{abundance.weighted}{Logical. TRUE calculates abundance-weighted mean pairwise phylogenetic distance. 
FALSE calculates presence-absence mean pairwise phylogenetic distance.}

\item{trim}{Logical. TRUE removes symbionts that occupy one host sample. FALSE keeps all symbionts.}
}
\value{
A data frame with symbiont identifiers and phylogenetic specificity values.
}
\description{
Calculate phylogenetic specificity not relativized by null mdoels.
}
\details{
Hosts are labeled by their sampling origin with a period and number identifier after their species name (e.g., hostA.1). This naming scheme is required and helps differentiate host samples that are of the same species. If this naming scheme does not apply to your experimental design, you can add in 'dummy variables' (e.g., .1, .2, .3, etc. after each host species or sample identifer).

Host specificity is calculated for a symbiont across the entire host community. For example, a symbiont found in given host would be evaluated for its host specificity relative to all hosts that are present in a given dataset.

More positive values indicate a narrower symbiont niche and thus higher host specificity. Structural and phylogenetic specificity were negated (multipled by -1) to make this consistent across all metrics.
}
\examples{
# Calculate mean pairwise phylogenetic distance per symbiont
\donttest{mpd <- phylogenetic.specificity(quad.rarefied, utree, abundance.weighted = TRUE, trim = TRUE)}
}
