% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_scripts.R
\name{relative.phylogenetic}
\alias{relative.phylogenetic}
\title{Deviance in Phylogenetic Specificity to Null Models}
\usage{
relative.phylogenetic(
  x,
  utree,
  null.model = c("taxa.labels", "richness", "frequency", "sample.pool",
    "phylogeny.pool", "independentswap", "trialswap"),
  iterations = 100,
  abundance.weighted = TRUE,
  model = c("first", "second"),
  trim = TRUE,
  notify = TRUE
)
}
\arguments{
\item{x}{Data frame. Host by symbiont data frame with hosts populating rows and symbionts populating columns.}

\item{utree}{Newick formatted host phylogenetic tree. Cophenetic distances will be calculated the function.}

\item{null.model}{Randomization method. Usage borrowed directly from picante::mpd 
Specify as "taxa.labels", "richness", "frequency", "sample.pool", "phylogeny.pool", "independentswap" or "trialswap".}

\item{iterations}{Integer. Indicate the number of randomized communities to generate. NOTE: This function can take some time.}

\item{abundance.weighted}{Logical. TRUE calculates abundance-weighted mean pairwise phylogenetic distance. 
FALSE calculates presence-absence mean pairwise phylogenetic distance.}

\item{model}{Character. Specify whether the null expectation should be approximated as a first-(linear) or second-(quadratic) order function.}

\item{trim}{Logical. TRUE removes symbionts that occupy one host species. FALSE keeps all symbionts.}

\item{notify}{Logical. TRUE prints the current iteration of the for loop.}
}
\value{
A list. First element in the list is a data frame with columns that refer to the host sample identifiers, mean relative phylogenetic specificity, standard error of phylogenetic specificities, number of symbionts per host sample and average symbiont read abundance. All subsequent elements of the list are plots of absolute phylogenetic specificity as a function of natural log symbiont read abundance with the null model in black (derived from host specificity of symbionts from randomized communities in grey) and the symbiont absolute phylogenetic specificities in red. Within graphs, there is an inset equation that refers to the null model expectation.
}
\description{
Calculate the deviance in observed phylogenetic specificity to a null model of structural specificity per symbiont and average output per sample.
}
\details{
Please make sure your host or sample identifiers match the labels on the host phylogenetic tree. 

Host specificity for a symbiont is evaluated across the entire host community. More positive values indicate a narrower symbiont niche and higher host specificity. 

Deviance calculations are measured per symbiont and averaged per host sample. The host specificities of each symbiont are averaged to calculate the mean host specificity for symbionts within a given host. Unlike relative structural, network and beta-specificity, relative phylogenetic specificity is not calculated as the vertical distance between absolute host specificity and the null model with respect to read abundance. Instead, relative phylogenetic specificity is calculated as the difference between absolute phylogenetic specificity and the mean value of phylogenetic specificity from randomized communities divided by the standard deviation of the null distribution of absolute phylogenetic specificity. 

For phylogenetic specificity, as the number of host species an symbiont occupies approaches the total number of host species in the community, the range of possible phylogenetic specificity values converges to a single value, or the MPD of the entire plant community. This produces a ‘funnel-shaped’ relationship between MPD and host species richness per symbiont and is often correlated with symbiont abundance. To account for this, relative phylogenetic specificity is calculated as the standardized effect size of MPD to quantify the difference between absolute phylogenetic specificity and the mean value of phylogenetic specificity from randomized communities.

A relative host specificity value greater than zero indicates that an symbiont was more host-specific relative to symbionts with the same read abundances within randomized communities. 

see documentation for `picante` for more documentation regarding the ses.mpd function
}
\examples{
# Calculate mean deviance per symbiont per host sample
\donttest{phy.dev <- relative.phylogenetic(quad.rarefied, utree, null.model = "taxa.labels", iterations = 100)}
\donttest{phy.dev[[1]]} # View data frame of output 
# Visualize null vs. absolute host specifities 
\donttest{phy.dev[[2]]} # View first graph 
\donttest{phy.dev[[81]]} # View last graph 
}
