% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity_scripts.R
\name{structural.specificity}
\alias{structural.specificity}
\title{Absolute Structural Specificity}
\usage{
structural.specificity(x, abundance.weighted = TRUE, trim = TRUE)
}
\arguments{
\item{x}{Data frame. Host by symbiont data frame with hosts populating rows and symbionts populating columns.}

\item{abundance.weighted}{Logical. TRUE calculates Shannon's H per symbiont. FALSE calculates host richness per symbiont.}

\item{trim}{Logical. TRUE removes symbionts that occupy one host species. FALSE keeps all symbionts.}
}
\value{
A data frame with symbiont identifiers and absolute structural specificity values.
}
\description{
Calculate absolute structural specificity not relativized by null mdoels.
}
\details{
Hosts are labeled by their species name with a period and number identifier (e.g., hostA.1) to differentiate host samples that are of the same species. This naming scheme is required because host specificity is quantified at the level of host species and not host samples.  If this naming scheme does not apply to your experimental design, you should still add  identifiers (e.g., .1, .2, .3, etc.) after each host species or sample identifer.

Host specificity for a symbiont is evaluated across the entire host community. More positive values indicate a narrower symbiont niche and higher host specificity. Structural specificity is negated (multipled by -1) to make this consistent across all metrics.
}
\examples{
# Calculate host species richness per symbiont
\donttest{host.richness <- structural.specificity(quad.rarefied, abundance.weighted = FALSE, trim = TRUE)}

# Calculate host species richness per symbiont
\donttest{shannons.h <- structural.specificity(quad.rarefied, abundance.weighted = TRUE, trim = TRUE)}
}
\references{
Austen Apigo and Ryoko Oono. 2021. Novel metrics reveal plant abundance, but not plant evolutionary history, shape host specificity in foliar fungal symbionts. In review.
}
