% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultrametric_tree.R
\docType{data}
\name{utree}
\alias{utree}
\title{Ultrametric tree for phylogenetic specificity analyses}
\format{
A phylogenetic tree 80 tips in newick format.
}
\source{
Apigo and Oono 2020
}
\usage{
data(utree)
}
\description{
Ultrametric phylogenetic tree used in Apigo and Oono 2020 to 
calculate phylogenetic specificity. This matrix includes 80 plant 
samples from 38 plant species.
}
\examples{
data(utree)

\donttest{structural.specificity(quad_rarefied)}
}
\references{
Apigo and Oono 2020.
}
\keyword{datasets}
